+++ src/java/florastore/deliveryOrder2/DeliverDTO.java
@@ -1,4 +1,4 @@
-package florastore.DeliveryOrder;
+package florastore.deliveryOrder2;
 
 import java.sql.Timestamp;
