+++ src/java/florastore/manageEvent2/CloseEventServlet.java
@@ -1,17 +1,12 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package florastore.DeliveryOrder;
+package florastore.manageEvent2;
 
+import florastore.account.AccountDTO;
+import florastore.event.EventDAO;
 import florastore.utils.MyAppConstants;
 import java.io.IOException;
 import java.sql.SQLException;
-import java.util.List;
 import java.util.Properties;
 import javax.naming.NamingException;
-import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -20,45 +15,33 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-@WebServlet(name = "MarkAsAcceptServlet", urlPatterns = {"/MarkAsAcceptServlet"})
-public class MarkAsAcceptServlet extends HttpServlet {
+/**
+ *
+ * @author ASUS
+ */
+@WebServlet(name = "CloseEventServlet", urlPatterns = {"/CloseEventServlet"})
+public class CloseEventServlet extends HttpServlet {
 
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
 
+        AccountDTO dto = null;
         ServletContext context = request.getServletContext();
         Properties siteMap = (Properties) context.getAttribute("SITE_MAP");
-        String url = (String) siteMap.get(MyAppConstants.Delivery.ERROR_PAGE);
-
+        String url = (String) siteMap.get(MyAppConstants.ManageEvent.ERROR_PAGE);
         HttpSession session = request.getSession();
-        String getEventOrderID = request.getParameter("getEventOrderID").trim();
-        String getEventOrderAmount = request.getParameter("getEventOrderAmount");
-        String getFullName = (String) session.getAttribute("USERNAME");
+        String EventID = request.getParameter("getEventID");
 
-        int eventOrderID = Integer.parseInt(getEventOrderID);
-        int staffID = (int) session.getAttribute("Staff_ID");
-
-        double staffBalance = (double) session.getAttribute("Staff_Balance");
-        double eventOrderAmount = Double.parseDouble(getEventOrderAmount);
-        double balance = staffBalance - eventOrderAmount;
         try {
-            if (balance >= 0) {
-                DeliverDAO dao = new DeliverDAO();
-                boolean result = dao.markAsGet(eventOrderID, staffID);
-                List<DeliverDTO> orderToDelivery = dao.getOrder(getFullName, staffID);          //lấy danh sách các đơn hàng để đi giao
-                session.setAttribute("Total_Order", orderToDelivery.size());
-                if (!result) {                                                      //đơn hàng đã bị người khác nhận
-                    request.setAttribute("FoundError", "Đơn hàng đã được nhận bởi nhân viên khác!");
-                } else {
-                    dao.setDeliveryStaffBalance(staffID, balance);
-                    session.removeAttribute("Staff_Balance");
-                    session.setAttribute("Staff_Balance", balance);
+            dto = (AccountDTO) session.getAttribute("USER");
+            if ("Admin".equals(dto.getRole())) {
+                EventDAO dao = new EventDAO();
+                boolean result = dao.closeEvent(EventID);
+                if (result) {
+                    url = (String) siteMap.get(MyAppConstants.ManageEvent.VIEW_EVENT);
                 }
-            } else {
-                request.setAttribute("FoundError2", "Số dư trong tài khoản của bạn không đủ để nhận đơn hàng này!");
             }
-            url = (String) siteMap.get(MyAppConstants.Delivery.SHIPPER_ORDER);
         } catch (SQLException ex) {
             log("EventServlet _SQL_ " + ex.getMessage());
         } catch (ClassNotFoundException ex) {
@@ -66,9 +49,7 @@ public class MarkAsAcceptServlet extends HttpServlet {
         } catch (NamingException ex) {
             log("EventServlet _Naming_ " + ex.getMessage());
         } finally {
-            RequestDispatcher rd = request.getRequestDispatcher(url);
-            rd.forward(request, response);
-
+            response.sendRedirect(url);
         }
     }
