+++ src/java/florastore/servlet/ViewOrderServlet.java
@@ -6,7 +6,7 @@
 package florastore.servlet;
 
 import com.google.gson.Gson;
-import florastore.ManageEvent.TotalPriceDTO;
+import florastore.manageEvent2.TotalPriceDTO;
 import florastore.account.AccountDTO;
 import florastore.event.EventDAO;
 import florastore.event.EventOrderDTO;
