+++ web/WEB-INF/siteMap.properties
@@ -133,3 +133,7 @@ taotaikhoan=CreateEWalletServlet
 eWallet=delivery/eWalletMachine.jsp
 sellerUpdateEventPage=seller/sellerUpdateEvent.jsp
 sellerUpdateEvent=sellerUpdateEventServlet
+sellerViewOrders=sellerViewOrdersServlet
+sellerOrderInformation=sellerOrderInformationServlet
+orderDeliveringPage=seller/orderDelivering.jsp
+sellerCancelOrder=MarkAsCancelServlet
