+++ web/seller/feedbackForSeller.jsp
@@ -66,6 +66,7 @@
                                 <a href=""><i class="ri-file-list-line"></i>Quản lý đơn hàng<i class="ri-add-circle-line"></i></a>
                                 <ul class="sub-menu">
                                     <li><a class="ri-arrow-right-s-fill" href="viewOrderAction">Đơn hàng chờ giao</a></li>
+                                    <li><a class="ri-arrow-right-s-fill" href="sellerViewOrders">Đơn hàng đang giao</a></li>
                                     <li><a class="ri-arrow-right-s-fill" href="viewDeliveredAction">Đơn hàng đã giao</a></li>
                                 </ul>
                             </li>
